Added our results
The results are saved in CSV files.
The utils folder contains a file with helper functions that summarize and visualize the data.
main.ipynb shows all the results in a notebook and shows the instructions how to run the code to obtain them.
Note that the MIMIC-III dataset requires credentialized access