Fix HMM data caching issue
The Hidden Markov Model data generator created the required data upon
the first invocation, then cached these files and loaded from them
during subsequent calls. However the file name used for caching
did not include any of the parameters used for data generation.
This was a problem because the HMM unit test generated a much smaller
number of samples that contain shorter signals, while the main HMM
experiment used more samples and longer signals.

If the user first executes the unit tests and not the main code then
s/he ends up unintentionally using the shorter version of the data.
This was the reason for having incorrect values for the HMM experiment
in the original version of the published paper, which was later corrected.

The HMM module now creates different cache files for different settings,
making sure no such errors happen. Currently the file name only includes
numeric settings, which was enough to avoid the error, though in the
future it could be expanded to also include non-numeric ones as well.